Add AxesChart abstract class and strip Chart.java of axis state
Introduce AxesChart<ST extends AxesChartStyler, S extends AxesChartSeries>
as a new abstract class between Chart and the 7 axes-based chart types.

Moved from Chart to AxesChart:
- Fields: axisPair, annotations, xAxisTitle, yAxisTitle, yAxisGroupTitleMap
- Methods: getXAxis/getYAxis/getAxisPair, getXAxisFormat/getYAxisFormat,
  coordinate conversion methods, setXAxisTitle/setYAxisTitle,
  getYAxisGroupTitle/setYAxisGroupTitle, setCustomX/YAxisTickLabelsFormatter,
  addAnnotation, getYAxisLeftWidth

Removed the 4 TODO comments and null guards from Chart.
