+++ xchart/src/main/java/org/knowm/xchart/internal/chartpart/Chart.java
@@ -4,18 +4,12 @@ import java.awt.Graphics2D;
 import java.awt.RenderingHints;
 import java.awt.Shape;
 import java.awt.geom.Rectangle2D;
-import java.text.DecimalFormat;
-import java.text.Format;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
-import java.util.function.Function;
 import org.knowm.xchart.ToolTipType;
 import org.knowm.xchart.internal.series.Series;
-import org.knowm.xchart.style.AxesChartStyler;
 import org.knowm.xchart.style.Styler;
 
 /** An XChart Chart */
@@ -24,13 +18,8 @@ public abstract class Chart<ST extends Styler, S extends Series> implements ICha
   protected final ST styler;
   protected final ChartTitle<ST, S> chartTitle;
   protected final Map<String, S> seriesMap = new LinkedHashMap<>();
-  protected final ArrayList<ChartPart> annotations = new ArrayList<>();
 
   /** Chart Parts */
-  // TODO maybe move this to a secondary abstract class for inheritors with axes. Pie charts don't
-  // have an axis for example
-  protected AxisPair<?, ?> axisPair;
-
   protected Plot_<ST, S> plot;
   protected Legend_<ST, S> legend;
 
@@ -39,13 +28,6 @@ public abstract class Chart<ST extends Styler, S extends Series> implements ICha
 
   private int height;
   private String title = "";
-  // TODO maybe move these to a secondary abstract class for inheritors with axes. Pie charts don't
-  // have an axis for example
-  private String xAxisTitle = "";
-  private String yAxisTitle = "";
-
-  // TODO Does this belong here for all chart types?
-  private final Map<Integer, String> yAxisGroupTitleMap = new HashMap<>();
 
   /**
    * Constructor
@@ -60,7 +42,6 @@ public abstract class Chart<ST extends Styler, S extends Series> implements ICha
     this.height = height;
     this.styler = styler;
 
-    // TODO move this out??
     this.chartTitle = new ChartTitle<ST, S>(this);
   }
 
@@ -126,68 +107,6 @@ public abstract class Chart<ST extends Styler, S extends Series> implements ICha
     this.title = title;
   }
 
-  public String getXAxisTitle() {
-
-    return xAxisTitle;
-  }
-
-  public void setXAxisTitle(String xAxisTitle) {
-
-    this.xAxisTitle = xAxisTitle;
-  }
-
-  public String getYAxisTitle() {
-
-    // TODO just call the getYAxisGroupTitle method passing in the 0th index
-    return yAxisTitle;
-  }
-
-  public void setYAxisTitle(String yAxisTitle) {
-
-    this.yAxisTitle = yAxisTitle;
-  }
-
-  // TODO these related methods don't make sense for all chart types
-  public String getYAxisGroupTitle(int yAxisGroup) {
-
-    String title = yAxisGroupTitleMap.get(yAxisGroup);
-    if (title == null) {
-      return yAxisTitle;
-    }
-    return title;
-  }
-
-  public void setYAxisGroupTitle(int yAxisGroup, String yAxisTitle) {
-
-    yAxisGroupTitleMap.put(yAxisGroup, yAxisTitle);
-  }
-
-  public void addAnnotation(Annotation annotation) {
-
-    annotations.add(annotation);
-    annotation.init(this);
-  }
-
-  /**
-   * @Deprecated - use styler instead
-   *
-   * @param customFormattingFunction
-   */
-  public void setCustomXAxisTickLabelsFormatter(Function<Double, String> customFormattingFunction) {
-    AxesChartStyler axesChartStyler = (AxesChartStyler) (styler);
-    axesChartStyler.setxAxisTickLabelsFormattingFunction(customFormattingFunction);
-  }
-
-  /**
-   * @Deprecated - use styler instead
-   *
-   * @param customFormattingFunction
-   */
-  public void setCustomYAxisTickLabelsFormatter(Function<Double, String> customFormattingFunction) {
-    AxesChartStyler axesChartStyler = (AxesChartStyler) (styler);
-    axesChartStyler.setyAxisTickLabelsFormattingFunction(customFormattingFunction);
-  }
-
   /** Chart Parts Getters */
   ChartTitle<ST, S> getChartTitle() {
 
@@ -202,105 +121,6 @@ public abstract class Chart<ST extends Styler, S extends Series> implements ICha
   Plot_<ST, S> getPlot() {
 
     return plot;
-  }
-
-  Axis_X<?, ?> getXAxis() {
-
-    return axisPair.getXAxis();
-  }
-
-  Axis_Y<?, ?> getYAxis() {
-
-    return axisPair.getYAxis();
-  }
-
-  Axis_Y<?, ?> getYAxis(int yIndex) {
-
-    return axisPair.getYAxis(yIndex);
-  }
-
-  AxisPair<?, ?> getAxisPair() {
-
-    return axisPair;
-  }
-
-  Format getXAxisFormat() {
-    return axisPair.getXAxis().getAxisTickCalculator().getAxisFormat();
-  }
-
-  Format getYAxisFormat() {
-    return axisPair.getYAxis().getAxisTickCalculator().getAxisFormat();
-  }
-
-  // TODO investigate this
-  Format getYAxisFormat(String yAxisDecimalPattern) {
-    final Format format;
-    if (yAxisDecimalPattern != null) {
-      format = new DecimalFormat(yAxisDecimalPattern);
-    } else {
-      format = axisPair.getYAxis().getAxisTickCalculator().getAxisFormat();
-    }
-    return format;
-  }
-
-  public double getChartXFromCoordinate(int screenX) {
-
-    if (axisPair == null) {
-      return Double.NaN;
-    }
-    return axisPair.getXAxis().getChartValue(screenX);
-  }
-
-  public double getChartYFromCoordinate(int screenY) {
-
-    if (axisPair == null) {
-      return Double.NaN;
-    }
-    return axisPair.getYAxis().getChartValue(screenY);
-  }
-
-  public double getChartYFromCoordinate(int screenY, int yIndex) {
-
-    if (axisPair == null) {
-      return Double.NaN;
-    }
-    return axisPair.getYAxis(yIndex).getChartValue(screenY);
-  }
-
-  public double getScreenXFromChart(double xValue) {
-
-    if (axisPair == null) {
-      return Double.NaN;
-    }
-    return axisPair.getXAxis().getScreenValue(xValue);
-  }
-
-  public double getScreenYFromChart(double yValue) {
-
-    if (axisPair == null) {
-      return Double.NaN;
-    }
-    return axisPair.getYAxis().getScreenValue(yValue);
-  }
-
-  public double getScreenYFromChart(double yValue, int yIndex) {
-
-    if (axisPair == null) {
-      return Double.NaN;
-    }
-    return axisPair.getYAxis(yIndex).getScreenValue(yValue);
-  }
-
-  //  ArrayList<ChartPart> getAnnotations() {
-  //
-  //    return annotations;
-  //  }
-
-  // TODO remove public
-  public double getYAxisLeftWidth() {
-
-    java.awt.geom.Rectangle2D.Double bounds = getAxisPair().getLeftYAxisBounds();
-    return bounds.width + bounds.x;
   }
 
   /**
